Add real "Header" object that handles multi-value management.
This is particularly useful for handling more esoteric cases. For example, `express` provides `WWW-Authenticate` _either_ as a single string value, _or_ an array of values. Deciding what action to take, or even finding the relevant value you're interested in can be a pain; this wrapper class deals with that issue.

It is waiting on `lodash.isMatch` to do its pattern matching, which is unfortunately only available in `^3.0.0` which hasn't landed in npm yet. What can ya do?