Modernize and remove `Header` class.
This is a fairly big adjustment, bringing everything up to date with ES6 syntax and the latest linting rules. However, the biggest change is that `Header` is no longer present and you must use `parse` and `format` yourself. This has the advantage of not requiring `lodash`, allowing the import of exactly the function(s) you need and avoiding any hardcoded logic for determining what header you're looking for.